[pfcwdorch] fix PfcWdSwOrch::doTask() starts WD action when WD wasn't started in warm boot
It appeared that pfc watchdog relied on a buggy behaviour of PortsOrch::allPortsReady().
In fixed PortsOrch::allPortsReady() you'll see that watchdog action is trying to start
before watchdog was started, because allPortsReady() in PfcWdOrch::doTask() returned false.
Before the fix watchdog was started before, because allPortsReady() lied that ports are ready
when they were not.

Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>