[RouteOrch] Publish route state for route to Loopback interface
Route to loopback interface is not advertised to peers when FIB suppression is enabled.

Route is queued:
```
admin@sonic:~$ show ipv6 route |grep q
       > - selected route, * - FIB route, q - queued route, r - rejected route
S>qfc00:1::/64 [1/0] is directly connected, Loopback0, 00:02:00
```

With this change:
```
admin@sonic:~$ show ipv6 route | grep q
       > - selected route, * - FIB route, q - queued route, r - rejected route
```

The proper way would be to program route to HW and then publish the
route state, however to keep the existing behaviour this W/A is proposed
until a final solution.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>